+++ .github/workflows/build-release-image.yaml
@@ -87,6 +87,7 @@ jobs:
 
       - name: Replace Variables
         if: ${{ contains(vars.OMNISTRATE_RELEASE_PLANS, matrix.plans.key) }}
+        working-directory: compose
         run: |
           sed -i 's/$GcpProjectId/${{ vars.GCP_PROJECT_ID }}/g' ${{ matrix.plans.file }}
           sed -i 's/$GcpProjectNumber/${{ vars.GCP_PROJECT_NUMBER }}/g' ${{ matrix.plans.file }}
@@ -102,6 +103,7 @@ jobs:
 
       - name: Remove e2-machine from pro and enterprise
         if: ${{ contains( vars.OMNISTRATE_RELEASE_PLANS, matrix.plans.key) && (matrix.plans.key == 'pro' || matrix.plans.key == 'enterprise') }}
+        working-directory: compose
         run: |
           sed -i 's/- e2-medium//g' ${{ matrix.plans.file }}
           sed -i 's/- t2.medium//g' ${{ matrix.plans.file }}
@@ -111,7 +113,7 @@ jobs:
         uses: actions/upload-artifact@v4 
         with:
           name: ${{ matrix.plans.file }}
-          path: ${{ matrix.plans.file }}
+          path: compose/${{ matrix.plans.file }}
 
       - name: Create FALKORDB_PASSWORD/ADMIN_PASSWORD files
         run: |
@@ -131,7 +133,7 @@ jobs:
         with:
           username: ${{ secrets.OMNISTRATE_USERNAME }}
           password: ${{ secrets.OMNISTRATE_PASSWORD }}
-          file: ${{ matrix.plans.file }}
+          file: compose/${{ matrix.plans.file }}
           service-name: ${{ matrix.plans.name }}
           environment: dev
           environment-type: dev
