+++ compose/omnistrate.free.yaml
