+++ compose/omnistrate.pro.yaml
