+++ compose/omnistrate.startup.yaml
