+++ scripts/cleanup-qa-instances-script.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python3
+
 import requests
 import time
 import os
