+++ tests/classes/omnistrate_types.py
