+++ tests/failover_sentinel.py
