+++ tests/test_replication_replicas.py
@@ -1,7 +1,12 @@
 import sys
 import signal
+from pathlib import Path
 from random import randbytes
-from pathlib import Path  # if you haven't already done so
+from redis import Redis
+from redis.backoff import ExponentialBackoff
+from redis.retry import Retry
+from redis.exceptions import ConnectionError, TimeoutError, ReadOnlyError
+import threading
 import socket
 
 file = Path(__file__).resolve()
@@ -15,15 +20,12 @@ with suppress(ValueError):
     sys.path.remove(str(parent))
 
 import logging
-
-logging.basicConfig(level=logging.DEBUG, format="%(asctime)s - %(message)s")
-
 import time
 import os
-from omnistrate_tests.classes.omnistrate_fleet_instance import OmnistrateFleetInstance
-from omnistrate_tests.classes.omnistrate_fleet_api import OmnistrateFleetAPI
+from tests.classes.omnistrate_fleet_instance import OmnistrateFleetInstance
+from tests.classes.omnistrate_fleet_api import OmnistrateFleetAPI
 import argparse
-from redis.exceptions import ResponseError, OutOfMemoryError
+
 parser = argparse.ArgumentParser()
 parser.add_argument("omnistrate_user")
 parser.add_argument("omnistrate_password")
@@ -47,12 +49,10 @@ parser.add_argument("--storage-size", required=False, default="30")
 parser.add_argument("--tls", action="store_true")
 parser.add_argument("--rdb-config", required=False, default="medium")
 parser.add_argument("--aof-config", required=False, default="always")
-parser.add_argument("--persist-instance-on-fail",action="store_true")
-parser.add_argument("--custom-network", required=False)
+parser.add_argument("--replica-count", required=False, default="2")
+parser.add_argument("--persist-instance-on-fail", action="store_true")
 parser.add_argument("--network-type", required=False, default="PUBLIC")
-parser.add_argument("--host-count", required=False, default="6")
-parser.add_argument("--cluster-replicas", required=False, default="1")
-parser.add_argument("--maxmemory", required=False, default="")
+parser.add_argument("--custom-network", required=False)
 
 parser.add_argument(
     "--deployment-create-timeout-seconds", required=False, default=2600, type=int
@@ -82,7 +82,7 @@ if not args.persist_instance_on_fail:
     signal.signal(signal.SIGINT, signal_handler)
 
 
-def test_out_of_memory():
+def test_add_remove_replica():
     global instance
 
     omnistrate = OmnistrateFleetAPI(
@@ -101,11 +101,9 @@ def test_out_of_memory():
     )
 
     logging.info(f"Product tier id: {product_tier.product_tier_id} for {args.ref_name}")
-
     network = None
     if args.custom_network:
         network = omnistrate.network(args.custom_network)
-
     instance = omnistrate.instance(
         service_id=args.service_id,
         service_provider_id=service.service_provider_id,
@@ -123,6 +121,7 @@ def test_out_of_memory():
     )
 
     try:
+        password = randbytes(16).hex()
         instance.create(
             wait_for_ready=True,
             deployment_cloud_provider=args.cloud_provider,
@@ -131,78 +130,188 @@ def test_out_of_memory():
             name=args.instance_name,
             description=args.instance_description,
             falkordb_user="falkordb",
-            falkordb_password=randbytes(16).hex(),
+            falkordb_password=password,
             nodeInstanceType=args.instance_type,
             storageSize=args.storage_size,
             enableTLS=args.tls,
             RDBPersistenceConfig=args.rdb_config,
             AOFPersistenceConfig=args.aof_config,
-            maxMemory=args.maxmemory,
-            custom_network_id=network.network_id if network else None,
+            custom_network_id=network.network_id if network else None
         )
-        
+
         try:
             ip = resolve_hostname(instance=instance)
-            logging.info(f"Instance endpoint {instance.get_cluster_endpoint()['endpoint']} resolved to {ip}")
+            logging.info(
+                f"Instance endpoint {instance.get_cluster_endpoint()['endpoint']} resolved to {ip}"
+            )
         except TimeoutError as e:
             logging.error(f"DNS resolution failed: {e}")
             raise Exception("Instance endpoint not ready: DNS resolution failed") from e
-        
-        # Stress test the instance
-        stress_test_out_of_memory(instance,resource_key=args.resource_key)
+
+        add_data(instance)
+
+        thread_signal = threading.Event()
+        error_signal = threading.Event()
+        thread = threading.Thread(
+            target=test_zero_downtime,
+            args=(thread_signal, error_signal, instance, args.tls),
+        )
+
+        thread.start()
+
+        check_data(instance)
+
+        change_replica_count(instance, int(args.replica_count) + 1)
+
+        test_fail_over(instance)
+
+        change_replica_count(instance, int(args.replica_count))
+
+        check_data(instance)
+
+        thread_signal.set()
+        thread.join()
+
     except Exception as e:
         logging.exception(e)
         if not args.persist_instance_on_fail:
-            instance.delete(network is not None)
+            instance.delete(False)
         raise e
 
     # Delete instance
-    instance.delete(network is not None)
+    instance.delete(False)
 
-    logging.info("Test passed")
+    if error_signal.is_set():
+        raise ValueError("Test failed")
+    else:
+        logging.info("Test passed")
 
 
+def change_replica_count(instance: OmnistrateFleetInstance, new_replica_count: int):
 
-def stress_test_out_of_memory(instance: OmnistrateFleetInstance,resource_key: str = None):
-    """Stress test the instance by allocating memory until it runs out of memory.
-    Should return this: '(error) OOM command not allowed when used memory > 'maxmemory''
-    Args:
-        instance: The OmnistrateFleetInstance to stress test
-    """
-    mediumQuery = "UNWIND RANGE(1, 100000) AS id CREATE (n:Person {name: 'Alice'})"
-    smallQuery = "UNWIND RANGE(1, 10000) AS id CREATE (n:Person {name: 'Alice'})"
+    logging.info(f"Changing replica count to {new_replica_count}")
+    instance.update_params(
+        numReplicas=new_replica_count,
+        wait_for_ready=True,
+    )
 
-    logging.info("Starting stress test")
+
+def test_fail_over(instance: OmnistrateFleetInstance):
+    logging.info("Testing failover to the newly created replica")
+
+    endpoint = instance.get_cluster_endpoint()
+    password = instance.falkordb_password
+    id_key = "sz" if args.resource_key == "single-Zone" else "mz"
+    retry = Retry(
+        ExponentialBackoff(base=1, cap=10),
+        retries=20,
+        supported_errors=(
+            TimeoutError,
+            ConnectionError,
+            ConnectionRefusedError,
+            ReadOnlyError,
+        ),
+    )
+    try:
+        client = Redis(
+            host=f"{endpoint['endpoint']}",
+            port=endpoint["ports"][0],
+            username="falkordb",
+            password=password,
+            decode_responses=True,
+            ssl=args.tls,
+            retry=retry,
+            retry_on_error=[
+                TimeoutError,
+                ConnectionError,
+                ConnectionRefusedError,
+                ReadOnlyError,
+            ],
+        )
+    except Exception as e:
+        logging.exception("Failed to connect to Sentinel!")
+        logging.info(e)
+
+    tout = time.time() + 600
+    while True:
+        if time.time() > tout:
+            raise Exception(f"Failed to failover to node-{id_key}-2")
+        try:
+            time.sleep(5)
+            print(client.execute_command("SENTINEL FAILOVER master"))
+            time.sleep(10)
+            master = client.execute_command("SENTINEL MASTER master")[3]
+            if master.startswith(f"node-{id_key}-2"):
+                break
+        except Exception as e:
+            logging.info(e)
+            continue
+    time.sleep(15)
+    check_data(instance)
+
+
+def add_data(instance: OmnistrateFleetInstance):
+    """This function should retrieve the instance host and port for connection, write some data to the DB, then check that the data is there"""
+    logging.info("Added data ....")
+    # Get instance host and port
     db = instance.create_connection(
         ssl=args.tls,
     )
+
     graph = db.select_graph("test")
 
-    if resource_key == 'free':
-        q = smallQuery
-    else:
-        q = mediumQuery
+    # Write some data to the DB
+    graph.query("CREATE (n:Person {name: 'Alice'})")
 
-    while True:
-        try:
-            response = graph.query(q)
-            logging.info(f"Response: {response}")
-        except Exception as e:
-            if isinstance(e, OutOfMemoryError) and 'memory' in str(e):
-                logging.info("Out of memory error detected")
-                break
-            logging.error(f"{e}")
-            logging.error("Failed to detect out of memory error")
-            raise e
-        
+    result = graph.query("MATCH (n:Person) RETURN n")
 
-def resolve_hostname(instance: OmnistrateFleetInstance,timeout=300, interval=1):
+    if len(result.result_set) == 0:
+        raise Exception("Data not found after write")
+
+
+def check_data(instance: OmnistrateFleetInstance):
+    logging.info("Retrieving data ....")
+    # Get instance host and port
+    db = instance.create_connection(ssl=args.tls, force_reconnect=True)
+
+    graph = db.select_graph("test")
+
+    result = graph.query("MATCH (n:Person) RETURN n")
+
+    if len(result.result_set) == 0:
+        raise Exception("Data did not persist after host count change")
+
+
+def test_zero_downtime(
+    thread_signal: threading.Event,
+    error_signal: threading.Event,
+    instance: OmnistrateFleetInstance,
+    ssl=False,
+):
+    """This function should test the ability to read and write while adding and removing a replica"""
+    try:
+        db = instance.create_connection(ssl=ssl, force_reconnect=True)
+
+        graph = db.select_graph("test")
+
+        while not thread_signal.is_set():
+            # Write some data to the DB
+            graph.query("CREATE (n:Person {name: 'Alice'})")
+            graph.ro_query("MATCH (n:Person {name: 'Alice'}) RETURN n")
+            time.sleep(3)
+    except Exception as e:
+        logging.exception(e)
+        error_signal.set()
+        raise e
+
+
+def resolve_hostname(instance: OmnistrateFleetInstance, timeout=300, interval=1):
     """Check if the instance's main endpoint is resolvable.
     Args:
         instance: The OmnistrateFleetInstance to check
         timeout: Maximum time in seconds to wait for resolution (default: 30)
         interval: Time in seconds between retry attempts (default: 1)
-    
+
     Returns:
         str: The resolved IP address
 
@@ -213,13 +322,13 @@ def resolve_hostname(instance: OmnistrateFleetInstance,timeout=300, interval=1):
     """
     if interval <= 0 or timeout <= 0:
         raise ValueError("Interval and timeout must be positive")
-    
+
     cluster_endpoint = instance.get_cluster_endpoint()
 
-    if not cluster_endpoint or 'endpoint' not in cluster_endpoint:
+    if not cluster_endpoint or "endpoint" not in cluster_endpoint:
         raise KeyError("Missing endpoint information in cluster configuration")
 
-    hostname = cluster_endpoint['endpoint']
+    hostname = cluster_endpoint["endpoint"]
     start_time = time.time()
 
     while time.time() - start_time < timeout:
@@ -229,8 +338,11 @@ def resolve_hostname(instance: OmnistrateFleetInstance,timeout=300, interval=1):
         except (socket.gaierror, socket.error) as e:
             logging.debug(f"DNS resolution attempt failed: {e}")
             time.sleep(interval)
-     
-    raise TimeoutError(f"Unable to resolve hostname '{hostname}' within {timeout} seconds.")
+
+    raise TimeoutError(
+        f"Unable to resolve hostname '{hostname}' within {timeout} seconds."
+    )
+
 
 if __name__ == "__main__":
-    test_out_of_memory()
+    test_add_remove_replica()
