+++ tests/test_upgrade_version.py
@@ -1,11 +1,11 @@
 import sys
 import signal
 from random import randbytes
-from pathlib import Path
+from pathlib import Path  # if you haven't already done so
 import threading
+# pylint: disable=no-name-in-module
+from utils import get_last_gh_tag
 import socket
-
-
 
 file = Path(__file__).resolve()
 parent, root = file.parent, file.parents[1]
@@ -23,9 +23,9 @@ logging.basicConfig(level=logging.DEBUG, format="%(asctime)s - %(message)s")
 
 import time
 import os
-from omnistrate_tests.classes.omnistrate_fleet_instance import OmnistrateFleetInstance
-from omnistrate_tests.classes.omnistrate_fleet_api import OmnistrateFleetAPI
-from omnistrate_tests.classes.falkordb_cluster import FalkorDBCluster
+from tests.classes.omnistrate_fleet_instance import OmnistrateFleetInstance
+from tests.classes.omnistrate_fleet_api import OmnistrateFleetAPI
+from tests.classes.omnistrate_types import TierVersionStatus
 import argparse
 
 parser = argparse.ArgumentParser()
@@ -41,11 +41,12 @@ parser.add_argument("--ref-name", required=False, default=os.getenv("REF_NAME"))
 parser.add_argument("--service-id", required=True)
 parser.add_argument("--environment-id", required=True)
 parser.add_argument("--resource-key", required=True)
-parser.add_argument("--replica-id", required=True)
 
 
 parser.add_argument("--instance-name", required=True)
-parser.add_argument("--instance-description", required=False, default="test-standalone")
+parser.add_argument(
+    "--instance-description", required=False, default="test-upgrade-version"
+)
 parser.add_argument("--instance-type", required=True)
 parser.add_argument("--storage-size", required=False, default="30")
 parser.add_argument("--tls", action="store_true")
@@ -53,11 +54,9 @@ parser.add_argument("--rdb-config", required=False, default="medium")
 parser.add_argument("--aof-config", required=False, default="always")
 parser.add_argument("--host-count", required=False, default="6")
 parser.add_argument("--cluster-replicas", required=False, default="1")
-
-parser.add_argument("--ensure-mz-distribution", action="store_true")
-parser.add_argument("--custom-network", required=False)
+parser.add_argument("--persist-instance-on-fail", action="store_true")
 parser.add_argument("--network-type", required=False, default="PUBLIC")
-
+parser.add_argument("--custom-network", required=False)
 
 parser.add_argument(
     "--deployment-create-timeout-seconds", required=False, default=2600, type=int
@@ -68,8 +67,6 @@ parser.add_argument(
 parser.add_argument(
     "--deployment-failover-timeout-seconds", required=False, default=2600, type=int
 )
-
-parser.add_argument("--persist-instance-on-fail",action="store_true")
 
 parser.set_defaults(tls=False)
 args = parser.parse_args()
@@ -89,8 +86,7 @@ if not args.persist_instance_on_fail:
     signal.signal(signal.SIGINT, signal_handler)
 
 
-
-def test_cluster():
+def test_upgrade_version():
     global instance
 
     omnistrate = OmnistrateFleetAPI(
@@ -110,10 +106,39 @@ def test_cluster():
 
     logging.info(f"Product tier id: {product_tier.product_tier_id} for {args.ref_name}")
 
+    # 1. List product tier versions
+    tiers = omnistrate.list_tier_versions(
+        service_id=args.service_id, tier_id=product_tier.product_tier_id
+    )
+
+    preferred_tier_tag = get_last_gh_tag()
+
+    preferred_tier = next(
+        (tier for tier in tiers if tier.description == preferred_tier_tag), None
+    )
+
+    if preferred_tier is None:
+        preferred_tier = next(
+            (tier for tier in tiers if tier.status == TierVersionStatus.PREFERRED), None
+        )
+
+    if preferred_tier is None:
+        raise ValueError("No preferred tier found")
+
+    last_tier = next(
+        (tier for tier in tiers if tier.status == TierVersionStatus.ACTIVE), None
+    )
+
+    if last_tier is None:
+        raise ValueError("No last tier found")
+
+    logging.info(f"Preferred tier: {preferred_tier.version}")
+    logging.info(f"Last tier: {last_tier.version}")
+
+    # 2. Create omnistrate instance with previous version
     network = None
     if args.custom_network:
         network = omnistrate.network(args.custom_network)
-
     instance = omnistrate.instance(
         service_id=args.service_id,
         service_provider_id=service.service_provider_id,
@@ -129,9 +154,7 @@ def test_cluster():
         deployment_delete_timeout_seconds=args.deployment_delete_timeout_seconds,
         deployment_failover_timeout_seconds=args.deployment_failover_timeout_seconds,
     )
-
     try:
-        password = randbytes(16).hex()
         instance.create(
             wait_for_ready=True,
             deployment_cloud_provider=args.cloud_provider,
@@ -140,7 +163,7 @@ def test_cluster():
             name=args.instance_name,
             description=args.instance_description,
             falkordb_user="falkordb",
-            falkordb_password=password,
+            falkordb_password=randbytes(16).hex(),
             nodeInstanceType=args.instance_type,
             storageSize=args.storage_size,
             enableTLS=args.tls,
@@ -148,189 +171,89 @@ def test_cluster():
             AOFPersistenceConfig=args.aof_config,
             hostCount=args.host_count,
             clusterReplicas=args.cluster_replicas,
-            custom_network_id=network.network_id if network else None,
+            product_tier_version=last_tier.version,
+            custom_network_id=network.network_id if network else None
 
         )
-        
+
         try:
             ip = resolve_hostname(instance=instance)
-            logging.info(f"Instance endpoint {instance.get_cluster_endpoint(network_type=args.network_type)['endpoint']} resolved to {ip}")
+            logging.info(f"Instance endpoint {instance.get_cluster_endpoint()['endpoint']} resolved to {ip}")
         except TimeoutError as e:
             logging.error(f"DNS resolution failed: {e}")
             raise Exception("Instance endpoint not ready: DNS resolution failed") from e
         
-        thread_signal = threading.Event()
-        error_signal = threading.Event()
-        thread = threading.Thread(
-            target=test_zero_downtime,
-            args=(thread_signal, error_signal, instance, args.tls),
+        # 3. Add data to the instance
+        add_data(instance)
+
+        thread_signal = None
+        error_signal = None
+        thread = None
+        if "standalone" not in args.instance_name:
+            thread_signal = threading.Event()
+            error_signal = threading.Event()
+            thread = threading.Thread(
+                target=test_zero_downtime,
+                args=(thread_signal, error_signal, instance, args.tls),
+            )
+            thread.start()
+
+        # 4. Upgrade version for the omnistrate instance
+        upgrade_timer = time.time()
+        instance.upgrade(
+            service_id=args.service_id,
+            product_tier_id=product_tier.product_tier_id,
+            source_version=last_tier.version,
+            target_version=preferred_tier.version,
+            wait_until_ready=True,
         )
-        thread.start()
 
-        if args.ensure_mz_distribution:
-            test_ensure_mz_distribution(instance, password)
+        if "standalone" not in args.instance_name:
+            thread_signal.set()
+            thread.join()
 
-        # Test failover and data loss
-        test_failover(instance)
+        logging.info(f"Upgrade time: {(time.time() - upgrade_timer):.2f}s")
 
-        # Wait for the zero_downtime
-        thread_signal.set()
-        thread.join()
-
-        # Test stop and start instance
-        test_stop_start(instance)
+        # 6. Verify the upgrade was successful
+        query_data(instance)
     except Exception as e:
         logging.exception(e)
         if not args.persist_instance_on_fail:
-            instance.delete(network is not None)
+            instance.delete(False)
         raise e
 
-    # Delete instance
-    instance.delete(network is not None)
+    # 7. Delete the instance
+    instance.delete(False)
 
-    if error_signal.is_set():
+    if "standalone" not in args.instance_name and error_signal.is_set():
         raise ValueError("Test failed")
     else:
         logging.info("Test passed")
 
 
-def test_ensure_mz_distribution(instance: OmnistrateFleetInstance, password: str):
-    """This function should ensure that each shard is distributed across multiple availability zones"""
+def add_data(instance: OmnistrateFleetInstance):
 
-    instance_details = instance.get_instance_details()
-    network_topology: dict = instance.get_network_topology(force_refresh=True)
-
-    params = (
-        instance_details["result_params"]
-        if "result_params" in instance_details
-        else None
-    )
-
-    # if not params:
-    #     raise Exception("No result_params found in instance details")
-
-    # host_count = int(params["hostCount"]) if "hostCount" in params else None
-
-    # if not host_count:
-    #     raise Exception("No hostCount found in instance details")
-
-    # cluster_replicas = (
-    #     int(params["clusterReplicas"]) if "clusterReplicas" in params else None
-    # )
-
-    # if not cluster_replicas:
-    #     raise Exception("No clusterReplicas found in instance details")
-
-    resource_key = next(
-        (k for [k, v] in network_topology.items() if v["resourceName"] == "cluster-mz"),
-        None,
-    )
-
-    resource = network_topology[resource_key]
-
-    nodes = resource["nodes"]
-
-    if len(nodes) == 0:
-        raise Exception("No nodes found in network topology")
-
-    if len(nodes) != 6:
-        raise Exception(
-            f"Host count does not match number of nodes. Current host count: {6}; Number of nodes: {len(nodes)}"
-        )
-
-    cluster = FalkorDBCluster(
-        host=resource["clusterEndpoint"],
-        port=resource["clusterPorts"][0],
-        username="falkordb",
-        password=password,
-        ssl=params["enableTLS"] == "true" if "enableTLS" in params else False,
-    )
-
-    groups = cluster.groups(1)
-
-    for group in groups:
-        group_azs = set()
-        for node in group:
-            omnistrateNode = next(
-                (n for n in nodes if n["endpoint"] == node.hostname), None
-            )
-            if not omnistrateNode:
-                logging.warning(f"Node {node.hostname} not found in network topology")
-                continue
-
-            group_azs.add(omnistrateNode["availabilityZone"])
-
-        if len(group_azs) == 1:
-            raise Exception(
-                "Group is not distributed across multiple availability zones"
-            )
-
-        logging.info(
-            f"Group {group} is distributed across availability zones {group_azs}"
-        )
-
-    logging.info("Shards are distributed across multiple availability zones")
-
-
-def test_failover(instance: OmnistrateFleetInstance):
-    """This function should retrieve the instance host and port for connection, write some data to the DB, then trigger a failover. After X seconds, the instance should be back online and data should have persisted"""
     # Get instance host and port
-    db = instance.create_connection(
-        ssl=args.tls,
-        network_type=args.network_type,
-    )
+    db = instance.create_connection(ssl=args.tls)
 
     graph = db.select_graph("test")
 
     # Write some data to the DB
     graph.query("CREATE (n:Person {name: 'Alice'})")
 
-    # Trigger failover
-    instance.trigger_failover(
-        replica_id=args.replica_id,
-        wait_for_ready=True,
-    )
-    
-    graph = db.select_graph("test")
 
-    result = graph.query("MATCH (n:Person) RETURN n")
-
-    if len(result.result_set) == 0:
-        raise Exception("Data lost after failover")
-
-    logging.info("Data persisted after failover")
-
-
-def test_stop_start(instance: OmnistrateFleetInstance):
-    """This function should stop the instance, check that it is stopped, then start it again and check that it is running"""
+def query_data(instance: OmnistrateFleetInstance):
 
     # Get instance host and port
-    db = instance.create_connection(
-        ssl=args.tls,
-        network_type=args.network_type,
-    )
+    db = instance.create_connection(ssl=args.tls)
 
     graph = db.select_graph("test")
 
-    # Write some data to the DB
-    graph.query("CREATE (n:Person {name: 'Alice'})")
-
-    logging.info("Stopping instance")
-
-    instance.stop(wait_for_ready=True)
-
-    logging.info("Instance stopped")
-
-    instance.start(wait_for_ready=True)
-
-    graph = db.select_graph("test")
-
-    result = graph.query("MATCH (n:Person) RETURN n")
+    # Get info
+    result = graph.query("MATCH (n:Person) RETURN n.name")
 
     if len(result.result_set) == 0:
-        raise Exception("Data lost after stop/start")
-
-    logging.info("Instance started")
+        raise ValueError("No data found in the graph after upgrade")
 
 
 def test_zero_downtime(
@@ -339,12 +262,12 @@ def test_zero_downtime(
     instance: OmnistrateFleetInstance,
     ssl=False,
 ):
-    """This function should test the ability to read and write while a failover happens"""
+    """This function should test the ability to read and write while an upgrade version happens"""
     try:
-        db = instance.create_connection(ssl=ssl, force_reconnect=True, network_type=args.network_type)
+        db = instance.create_connection(ssl=ssl, force_reconnect=True)
 
         graph = db.select_graph("test")
-        
+
         while not thread_signal.is_set():
             # Write some data to the DB
             graph.query("CREATE (n:Person {name: 'Alice'})")
@@ -354,7 +277,7 @@ def test_zero_downtime(
         logging.exception(e)
         error_signal.set()
         raise e
-    
+
 def resolve_hostname(instance: OmnistrateFleetInstance,timeout=300, interval=1):
     """Check if the instance's main endpoint is resolvable.
     Args:
@@ -373,7 +296,7 @@ def resolve_hostname(instance: OmnistrateFleetInstance,timeout=300, interval=1):
     if interval <= 0 or timeout <= 0:
         raise ValueError("Interval and timeout must be positive")
     
-    cluster_endpoint = instance.get_cluster_endpoint(network_type=args.network_type)
+    cluster_endpoint = instance.get_cluster_endpoint()
 
     if not cluster_endpoint or 'endpoint' not in cluster_endpoint:
         raise KeyError("Missing endpoint information in cluster configuration")
@@ -392,4 +315,4 @@ def resolve_hostname(instance: OmnistrateFleetInstance,timeout=300, interval=1):
     raise TimeoutError(f"Unable to resolve hostname '{hostname}' within {timeout} seconds.")
 
 if __name__ == "__main__":
-    test_cluster()
+    test_upgrade_version()
